requirements.txt
asarPy==1.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
Flask==3.1.1
flatbuffers==25.2.10
humanfriendly==10.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
PyMatting==1.1.14
referencing==0.36.2
rembg==2.0.67
requests==2.32.4
rpds-py==0.26.0
scikit-image==0.25.2
scipy==1.16.0
sympy==1.14.0
tifffile==2025.6.11
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
